Fix build failures with LLVM-GCC 4.2 on Mac.
R=titzer@chromium.org

Review URL: https://codereview.chromium.org/431613004

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@22716 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/common-operator.h
@@ -24,7 +24,8 @@ class ControlOperator : public Operator1<int> {
  public:
   ControlOperator(IrOpcode::Value opcode, uint16_t properties, int inputs,
                   int outputs, int controls, const char* mnemonic)
-      : Operator1(opcode, properties, inputs, outputs, mnemonic, controls) {}
+      : Operator1(static_cast<uint8_t>(opcode), properties, inputs, outputs,
+                  mnemonic, controls) {}
 
   virtual OStream& PrintParameter(OStream& os) const { return os; }  // NOLINT
   int ControlInputCount() const { return parameter(); }
@@ -33,9 +34,9 @@ class ControlOperator : public Operator1<int> {
 class CallOperator : public Operator1<CallDescriptor*> {
  public:
   CallOperator(CallDescriptor* descriptor, const char* mnemonic)
-      : Operator1(IrOpcode::kCall, descriptor->properties(),
-                  descriptor->InputCount(), descriptor->ReturnCount(), mnemonic,
-                  descriptor) {}
+      : Operator1(static_cast<uint8_t>(IrOpcode::kCall),
+                  descriptor->properties(), descriptor->InputCount(),
+                  descriptor->ReturnCount(), mnemonic, descriptor) {}
 
   virtual OStream& PrintParameter(OStream& os) const {  // NOLINT
     return os << "[" << *parameter() << "]";
